% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardigans.R
\name{cardigan}
\alias{cardigan}
\title{Generates palettes based on Mr. Rogers cardigans}
\usage{
cardigan(n, type = c("discrete", "continuous"))
}
\arguments{
\item{n}{Number of colors desired. Default is to use all colors.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colors.
}
\description{
These color palettes are based on Owen Phillips blog post on 
[Mr. Rogers cardigans](https://theawl.com/every-color-of-cardigan-mister-rogers-wore-from-1979-2001-83c1faba2677) 
and [@Henryjean's code](https://github.com/Henryjean/Rogers-Cardigans).  
And the real work in getting these colors is from [Tim Lybarger](http://www.neighborhoodarchive.com/about/).
}
\examples{

}
\keyword{colors}
